refactor: update map-related interfaces and improve address handling logic
