+++ src/app/maps/interfaces/click-event.interface.ts
@@ -1,4 +1,4 @@
-import { MapCoordinate } from './coordinate.interface';
+import { MapCoordinate } from './map-coordinate.interface';
 
 export interface MapClickEvent {
   coordinate: MapCoordinate;
