+++ src/app/maps/interfaces/map-coordinate.interface.ts
